Remove ElectricDatabase for node-postgres and directly electrify the PG connection
